+++ fin/ChangeLog.md
@@ -1,6 +1,8 @@
 ## 0.2.1
 
 - Add `boring` instances
+- Explicitly implement `>=` and `>` for `Nat`.
+- `<=`, `>=` and `min` for `Nat` are lazier
 
 ## 0.2